update to image name
Fix non-existing resource on aro_delete

Increase redeployvm wait to 15 minutes

It was recently increased from 5 to 10 minutes but it isn't enough.  Per
documentation (linked in code now) it can take up to 15 minutes to ship
activity logs in some situations.

api: add vm types to current api

quota: add vm types to current api

api: add vm types validator

Add @m1kola

Make the admin portal REST API not perform redirects if user is not authenticated, introduce new API for that and reorganise slightly

Fixup for infraID and legacy retrofit (#1536)

* add InfraID
* fixup old clusters with legacy infraID

User RP credentials in DB script (#1559)

Add MSI back into prodfromenv (#1542)

add InfraID to admin API (#1560)

allow git branches ending in ".go"

Extend azure.Environment

use AROEnvironment everywhere

remove now-redundant validateCloudEnvironment

Centralize logic for extra cloud-aware domains

Make hardcoded domains cloud-aware

add cmd to trigger adminupdate against a cluster

Apply suggestions from code review


Fix aro operator not clearing empty but set volumeClaimTemplates in monitoring configmap

Consolidate monitoring controllers

Now check that CRG managedBy == openshiftcluster.id, if not, we do not run full delete

always delete role assignments/definitions

dont return err if we fail to get the resource group, continue with delete logic

fix warnf fmt

Fixes dev cluster deletion

change SP for prod

Makes pyenv target more platform agnostic

Make CI/E2E clusters public

Adj poller to check for payload

Fixup routefix for 4.7.15+

cleanup fix

RG flake

Add spew for delete errors

bump install versions

regen assets

remove routefix e2e test

Vendor update

API: disk encryption and encryption at host

Backend: disk encryption and encryption at host

make client

Unit test improvements

* Moves gomock.NewController(t) call into subtests, where possible
* Adds missing t.Run(tt.name, ...) calls

Enable RG tagging

Improves SKU request

enable failure gathering in dev

fix tests

Ci fixup

Fixes copy-paste issue

fixup fluentbit agent

Revert "Merge pull request #1591 from mjudeikis/fleuntbit.pin"

This reverts commit 57e4bf000d38df650b29247319f5796e0280ed32, reversing
changes made to 6108dc6b82dce88d8b8c11ff75401cf32f8d82ae.

+++ pkg/cluster/delete.go
@@ -14,6 +14,7 @@ import (
 	mgmtfeatures "github.com/Azure/azure-sdk-for-go/services/resources/mgmt/2019-07-01/features"
 	"github.com/Azure/go-autorest/autorest"
 	"github.com/Azure/go-autorest/autorest/azure"
+	"github.com/davecgh/go-spew/spew"
 
 	"github.com/Azure/ARO-RP/pkg/api"
 	"github.com/Azure/ARO-RP/pkg/env"
@@ -259,9 +260,21 @@ func (m *manager) deleteRoleDefinition(ctx context.Context) error {
 
 func (m *manager) Delete(ctx context.Context) error {
 	resourceGroup := stringutils.LastTokenByte(m.doc.OpenShiftCluster.Properties.ClusterProfile.ResourceGroupID, '/')
+	//In edge case of CRG not being managedBy ARO, we have a different delete path
+	//we will assume normal case and set rgManagedByARO to true CRG is managedby ARO
+	rgManagedByARO := true
+	rg, err := m.resourceGroups.Get(ctx, resourceGroup)
+	if err != nil {
+		m.log.Warnf("failed to get resourceGroup %s", err)
+	} else if !m.env.IsLocalDevelopmentMode() {
+		if rg.ManagedBy == nil || *rg.ManagedBy == "" || !strings.EqualFold(*rg.ManagedBy, m.doc.OpenShiftCluster.ID) {
+			rgManagedByARO = false
+			m.log.Infof("cluster resource group not managed by aro %s", *rg.Name)
+		}
+	}
 
 	m.log.Printf("deleting dns")
-	err := m.dns.Delete(ctx, m.doc.OpenShiftCluster)
+	err = m.dns.Delete(ctx, m.doc.OpenShiftCluster)
 	if err != nil {
 		return err
 	}
@@ -284,25 +297,35 @@ func (m *manager) Delete(ctx context.Context) error {
 		return err
 	}
 
-	m.log.Printf("deleting resources")
-	err = m.deleteResources(ctx)
-	if err != nil {
-		return err
-	}
+	// only delete if managedByARO
+	if rgManagedByARO {
+		m.log.Printf("deleting resources")
+		err = m.deleteResources(ctx)
+		if err != nil {
+			return err
+		}
 
-	m.log.Printf("deleting resource group %s", resourceGroup)
-	err = m.resourceGroups.DeleteAndWait(ctx, resourceGroup)
-	if detailedErr, ok := err.(autorest.DetailedError); ok &&
-		(detailedErr.StatusCode == http.StatusForbidden || detailedErr.StatusCode == http.StatusNotFound) {
-		err = nil
+		m.log.Printf("deleting resource group %s", resourceGroup)
+		err = m.resourceGroups.DeleteAndWait(ctx, resourceGroup)
+		// TODO(mjudeikis): Remove this once we know all the error flavors this is
+		// returning so we can unify checks bellow
+		if err != nil {
+			m.log.Printf("delete resource group failed: %s", spew.Sdump(err))
+		}
+		if detailedErr, ok := err.(autorest.DetailedError); ok &&
+			(detailedErr.StatusCode == http.StatusForbidden || detailedErr.StatusCode == http.StatusNotFound) {
+			err = nil
+		}
+		if azureerrors.HasAuthorizationFailedError(err) {
+			err = nil
+		}
+		if azureerrors.ResourceGroupNotFound(err) {
+			err = nil
+		}
+		if err != nil {
+			return err
+		}
 	}
-	if azureerrors.HasAuthorizationFailedError(err) {
-		err = nil
-	}
-	if err != nil {
-		return err
-	}
-
 	if !m.env.FeatureIsSet(env.FeatureDisableSignedCertificates) {
 		managedDomain, err := dns.ManagedDomain(m.env, m.doc.OpenShiftCluster.Properties.ClusterProfile.Domain)
 		if err != nil {
